requirements.txt
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0 
langgraph
langchain
langchain-google-genai
faiss-cpu
langchain-huggingface
gradio
spacy
numpy>=1.26.0